Succesfully refactored evaluate phase and files in examples directory
