+++ midca/examples/_baxter/homography.py
