+++ midca/examples/baxter_3d_camera.py
@@ -5,15 +5,16 @@ from midca.modules.plan import AsynchPyhopPlanner
 from midca.modules.intend import SimpleIntend
 from midca.modules.act import AsynchronousAct
 from midca.modules.interpret import InstructionReceiver
-from MIDCA.modules import simulator, evaluate
+from midca.modules.evaluate import EvalPointingFromFeedback
+from MIDCA.modules import simulator
 from MIDCA.modules._plan.asynch import asynch, operators_sr, methods_sr, monitors_sr
 from MIDCA.logging import Logger
 import inspect, os
 import rospy
 from std_msgs.msg import String
-from MIDCA.examples import Calibrate
+from MIDCA.examples._baxter import Calibrate
 from geometry_msgs.msg import Point, PointStamped
-from .baxter import *
+from midca.examples._baxter.baxter import *
 
 
 def ros_style_midca():
@@ -23,7 +24,7 @@ def ros_style_midca():
 
 	myMidca.append_module("Perceive", ROSObserver.ROSObserver())
 	myMidca.append_module("Interpret", InstructionReceiver.InstructionReceiver_sr())
-	myMidca.append_module("Eval", evaluate.EvalPointingFromFeedback())
+	myMidca.append_module("Eval", EvalPointingFromFeedback.EvalPointingFromFeedback())
 	myMidca.append_module("Intend", SimpleIntend.SimpleIntend())
 	myMidca.append_module("Plan", AsynchPyhopPlanner.AsynchPyhopPlanner_3d_camera(methods_sr.declare_methods,
 	operators_sr.declare_ops,monitors_sr.declare_monitors
