+++ midca/examples/cogsci_demo.py
@@ -1,11 +1,12 @@
 #!/usr/bin/env python
 from midca import base
-from midca.modules import simulator, evaluate
+from midca.modules import simulator
 from midca.modules.perceive import PerfectObserver
 from midca.modules.plan import PyHopPlanner
 from midca.modules.intend import SimpleIntend
 from midca.modules.interpret import ADistanceAnomalyNoter, TFStack, TFFire, ReactiveApprehend, UserGoalInput
 from midca.modules.act import SimpleAct
+from midca.modules.evaluate import SimpleEval, Scorer
 from midca.worldsim import domainread, stateread
 import inspect, os
 
@@ -46,7 +47,7 @@ myMidca.append_module("Simulate", simulator.ASCIIWorldViewer(display=DISPLAY_FUN
 myMidca.append_module("Perceive", PerfectObserver.PerfectObserver())
 myMidca.append_module("Interpret", ADistanceAnomalyNoter.ADistanceAnomalyNoter())
 #myMidca.append_module("Interpret", UserGoalInput.UserGoalInput())
-myMidca.append_module("Eval", evaluate.SimpleEval())
+myMidca.append_module("Eval", SimpleEval.SimpleEval())
 myMidca.append_module("Intend", SimpleIntend.SimpleIntend())
 myMidca.append_module("Plan", PyHopPlanner.PyHopPlanner(util.pyhop_state_from_world,
                                                     util.pyhop_tasks_from_goals,
@@ -59,7 +60,7 @@ myMidca.insert_module('Simulate', simulator.FireReset(), 0)
 myMidca.insert_module('Interpret', TFStack.TFStack(), 1)
 myMidca.insert_module('Interpret', TFFire.TFFire(), 2)
 myMidca.insert_module('Interpret', ReactiveApprehend.ReactiveApprehend(), 3)
-myMidca.insert_module('Eval', evaluate.Scorer(), 1) # this needs to be a 1 so that Scorer happens AFTER SimpleEval
+myMidca.insert_module('Eval', Scorer.Scorer(), 1) # this needs to be a 1 so that Scorer happens AFTER SimpleEval
 
 
 #tells the PhaseManager to copy and store MIDCA states so they can be accessed later.
