+++ midca/examples/cogsci_demo_mortar_goal_transforms.py
@@ -7,8 +7,9 @@ from midca.modules.intend import SimpleIntend
 from midca.modules.act import SimpleAct
 from midca.modules.interpret import TFFire,\
     SimpleMortarGoalGen, ReactiveApprehend, UserGoalInput, ADistanceAnomalyNoter
+from midca.modules.evaluate import SimpleEval, MortarScorer
 from midca.worldsim import domainread, stateread
-from midca.modules import simulator, evaluate
+from midca.modules import simulator
 from midca.metamodules import monitor, control, interpret, metaintend,  plan
 from midca.modules.gens import goaltransform
 from midca import base
@@ -67,7 +68,7 @@ myMidca.append_module("Simulate", simulator.ASCIIWorldViewer(display=DISPLAY_FUN
 myMidca.append_module("Perceive", PerfectObserver.PerfectObserver())
 myMidca.append_module("Interpret", ADistanceAnomalyNoter.ADistanceAnomalyNoter())
 #myMidca.append_module("Interpret", guide.UserGoalInput())
-myMidca.append_module("Eval", evaluate.SimpleEval())
+myMidca.append_module("Eval", SimpleEval.SimpleEval())
 myMidca.append_module("Intend", SimpleIntend.SimpleIntend())
 myMidca.append_module("Plan", PyHopPlanner.PyHopPlanner(util.mortar_pyhop_state_from_world,
                                                     util.mortar_pyhop_tasks_from_goals,
@@ -81,7 +82,7 @@ myMidca.append_module("Act", SimpleAct.SimpleAct())
 myMidca.insert_module('Interpret', SimpleMortarGoalGen.SimpleMortarGoalGen(), 1)
 #myMidca.insert_module('Interpret', guide.TFFire(), 2)
 myMidca.insert_module('Interpret', ReactiveApprehend.ReactiveApprehend(), 3)
-myMidca.insert_module('Eval', evaluate.MortarScorer(), 1) # this needs to be a 1 so that Scorer happens AFTER SimpleEval
+myMidca.insert_module('Eval', MortarScorer.MortarScorer(), 1) # this needs to be a 1 so that Scorer happens AFTER SimpleEval
 
 # add meta layer phases
 #for phase in ["Monitor", "Interpret", "Eval", "Intend", "Plan", "Control"]:
