+++ midca/examples/predicateworld.py
@@ -6,7 +6,8 @@ from midca.modules.plan import PyHopPlanner
 from midca.modules.intend import SimpleIntend
 from midca.modules.act import SimpleAct
 from midca.modules.interpret import  UserGoalInput, ADistanceAnomalyNoter
-from midca.modules import simulator, evaluate
+from midca.modules.evaluate import SimpleEval
+from midca.modules import simulator
 
 
 def UserGoalsMidca(domainFile, stateFile, display=print, goalsFile = None, argsPyHopPlanner=[]):
@@ -24,7 +25,7 @@ def UserGoalsMidca(domainFile, stateFile, display=print, goalsFile = None, argsP
     myMidca.append_module("Perceive", PerfectObserver.PerfectObserver())
     myMidca.append_module("Interpret", ADistanceAnomalyNoter.ADistanceAnomalyNoter())
     #myMidca.append_module("Interpret", UserGoalInput.UserGoalInput())
-    myMidca.append_module("Eval", evaluate.SimpleEval())
+    myMidca.append_module("Eval", SimpleEval.SimpleEval())
     myMidca.append_module("Intend", SimpleIntend.SimpleIntend())
     myMidca.append_module("Plan", PyHopPlanner.PyHopPlanner(*argsPyHopPlanner))
     myMidca.append_module("Act", SimpleAct.SimpleAct())
