+++ midca/examples/simple_run_arson.py
@@ -5,7 +5,8 @@ from midca.modules.plan import PyHopPlanner
 from midca.modules.intend import SimpleIntend
 from midca.modules.act import SimpleAct
 from midca.modules.interpret import ADistanceAnomalyNoter, TFStack, TFFire, UserGoalInput
-from midca.modules import simulator, evaluate
+from midca.modules.evaluate import SimpleEval
+from midca.modules import simulator
 from midca.worldsim import stateread, domainread
 import inspect, os
 
@@ -41,7 +42,7 @@ myMidca.append_module("Simulate", simulator.ASCIIWorldViewer())
 myMidca.append_module("Perceive", PerfectObserver.PerfectObserver())
 myMidca.append_module("Interpret", ADistanceAnomalyNoter.ADistanceAnomalyNoter())
 #myMidca.append_module("Interpret", UserGoalInput.UserGoalInput())
-myMidca.append_module("Eval", evaluate.SimpleEval())
+myMidca.append_module("Eval", SimpleEval.SimpleEval())
 myMidca.append_module("Intend", SimpleIntend.SimpleIntend())
 myMidca.append_module("Plan", PyHopPlanner.PyHopPlanner(util.pyhop_state_from_world,
                                                     util.pyhop_tasks_from_goals,
