+++ midca/examples/simple_run_arson_broken_shop_methods.py
@@ -5,7 +5,8 @@ from midca.modules.plan import PyHopPlannerBroken
 from midca.modules.intend import SimpleIntend
 from midca.modules.act import SimpleAct
 from midca.modules.interpret import ADistanceAnomalyNoter, TFStack, TFFire, UserGoalInput
-from midca.modules import simulator, evaluate
+from midca.modules.evaluate import SimpleEval
+from midca.modules import simulator
 from midca.metamodules import monitor, interpret, metaeval, metaintend, plan, control
 from midca.worldsim import domainread, stateread
 from midca.domains.blocksworld import blockstate, scene, util
@@ -45,7 +46,7 @@ myMidca.append_module("Simulate", simulator.ASCIIWorldViewer())
 myMidca.append_module("Perceive", PerfectObserver.PerfectObserver())
 myMidca.append_module("Interpret", ADistanceAnomalyNoter.ADistanceAnomalyNoter())
 myMidca.append_module("Interpret", UserGoalInput.UserGoalInput())
-myMidca.append_module("Eval", evaluate.SimpleEval())
+myMidca.append_module("Eval", SimpleEval.SimpleEval())
 myMidca.append_module("Intend", SimpleIntend.SimpleIntend())
 myMidca.append_module("Plan", PyHopPlannerBroken.PyHopPlannerBroken(util.pyhop_state_from_world,
                                                                 util.pyhop_tasks_from_goals,
