+++ midca/examples/simple_run_jshop.py
@@ -5,7 +5,8 @@ from midca.modules.plan import JSHOPPlanner
 from midca.modules.intend import SimpleIntend
 from midca.modules.act import SimpleAct
 from midca.modules.interpret import ADistanceAnomalyNoter, TFStack, TFFire, UserGoalInput, ReactiveApprehend
-from midca.modules import simulator, evaluate
+from midca.modules.evaluate import SimpleEval, Scorer
+from midca.modules import simulator
 
 from midca.worldsim import domainread, stateread
 import inspect, os
@@ -63,7 +64,7 @@ myMidca.append_module("Simulate", simulator.ASCIIWorldViewer(display=DISPLAY_FUN
 myMidca.append_module("Perceive", PerfectObserver.PerfectObserver())
 myMidca.append_module("Interpret", ADistanceAnomalyNoter.ADistanceAnomalyNoter())
 #myMidca.append_module("Interpret", UserGoalInput.UserGoalInput())
-myMidca.append_module("Eval", evaluate.SimpleEval())
+myMidca.append_module("Eval", SimpleEval.SimpleEval())
 myMidca.append_module("Intend", SimpleIntend.SimpleIntend())
 myMidca.append_module("Plan", JSHOPPlanner.JSHOPPlanner(util.jshop_state_from_world,
                                                     util.jshop_tasks_from_goals,
@@ -77,7 +78,7 @@ myMidca.insert_module('Simulate', simulator.FireReset(), 0)
 myMidca.insert_module('Interpret', TFStack.TFStack(), 1)
 myMidca.insert_module('Interpret', TFFire.TFFire(), 2)
 myMidca.insert_module('Interpret', ReactiveApprehend.ReactiveApprehend(), 3)
-myMidca.insert_module('Eval', evaluate.Scorer(), 1) # this needs to be a 1 so that Scorer happens AFTER SimpleEval
+myMidca.insert_module('Eval', Scorer.Scorer(), 1) # this needs to be a 1 so that Scorer happens AFTER SimpleEval
 
 
 #tells the PhaseManager to copy and store MIDCA states so they can be accessed later.
