+++ setup.py
@@ -38,6 +38,8 @@ setup(name='midca',
                 'midca.modules.intend',
                 'midca.modules.interpret',
                 'midca.modules.interpret._goalgen',
+                'midca.modules.interpret._xp_goal',
+                'midca.modules.interpret._adist',
                 'midca.vision',
                 'midca.experimental',
                 'midca.experimental.baxter',
